backend/requirements.txt
modal>=0.60.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
openai-whisper>=20231117
pyttsx3>=2.90
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
